+++ E3/validator.py
@@ -3,7 +3,7 @@ import re
 import signal
 
 from E3.utils import *
-from subprocess import Popen, PIPE
+from subprocess import Popen, PIPE, SubprocessError
 
 
 class Validator:
@@ -11,19 +11,20 @@ class Validator:
         self.tmp_path = tmp_path
         os.makedirs(self.tmp_path, exist_ok=True)
 
-    def validate(self, expression, instanceName):
-        tmpFile = os.path.join(self.tmp_path, instanceName + ".lean")
-        os.makedirs(os.path.dirname(tmpFile), exist_ok=True)
+    def validate(self, expression, instance_name):
+        tmp_file = os.path.join(self.tmp_path, instance_name + ".lean")
+        os.makedirs(os.path.dirname(tmp_file), exist_ok=True)
 
-        leanFile = format_test_file(expression)
-        with open(tmpFile, "w") as file:
-            file.write(leanFile)
+        lean_file = format_test_file(expression)
+        with open(tmp_file, "w") as file:
+            file.write(lean_file)
 
-        command = ["lake", "env", "lean", "--run", tmpFile]
-        process = Popen(
-            command, stdin=PIPE, stdout=PIPE, cwd=ROOT_DIR, preexec_fn=os.setsid
-        )
+        process = None
+        command = ["lake", "env", "lean", "--run", tmp_file]
         try:
+            process = Popen(
+                command, stdin=PIPE, stdout=PIPE, cwd=ROOT_DIR, preexec_fn=os.setsid
+            )
             stdout, _ = process.communicate()
             if stdout == b"":
                 return None
@@ -31,6 +32,8 @@ class Validator:
                 error = stdout.decode()
                 error = re.sub(r"/[^:]+:\d+:\d+: ", "", error)
                 return error
-        except:
-            os.killpg(os.getpgid(process.pid), signal.SIGTERM)
+        except (SubprocessError, OSError) as e:
+            print(f"Unexpected error: {e}")
+            if process:
+                os.killpg(os.getpgid(process.pid), signal.SIGTERM)
             return "Unexpected error"
